# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PolicyBot API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface PolicyQueryBody {
  /** The natural language policy question */
  query: string;
  /** Language of the query (en, hi, kn, etc.) */
  language?: string;
  /** Optional user persona (farmer, student, entrepreneur, etc.) */
  persona?: string;
}

export interface PolicyCompareBody {
  /**
   * List of scheme names to compare
   * @minItems 2
   */
  schemes: string[];
  /** Optional specific aspect to compare */
  query?: string;
}

export interface PolicyRecommendBody {
  /** User description or persona (e.g., "small farmer in Karnataka", "college student looking for scholarship") */
  persona: string;
  /** Areas of interest */
  interests?: string[];
}

export interface CitedChunk {
  text: string;
  scheme: string;
  source: string;
  relevanceScore: number;
}

export interface PolicyResponse {
  summary: string;
  eligibility?: string;
  benefits?: string;
  applicationProcess?: string;
  sources: CitedChunk[];
  confidenceScore: number;
  language?: string;
  noDataFound: boolean;
}

export type ComparisonRowValues = { [key: string]: string };

export interface ComparisonRow {
  aspect: string;
  values: ComparisonRowValues;
}

export interface PolicyComparisonResponse {
  schemes: string[];
  comparisonRows: ComparisonRow[];
  summary: string;
  sources: CitedChunk[];
  confidenceScore: number;
}

export interface RecommendedScheme {
  name: string;
  sector: string;
  summary: string;
  matchReason: string;
  eligibilityHighlight?: string;
  benefitHighlight?: string;
}

export interface PolicyRecommendationResponse {
  persona: string;
  recommendations: RecommendedScheme[];
  confidenceScore: number;
}

export interface SchemeEntry {
  id: string;
  name: string;
  sector: string;
  ministry: string;
  description: string;
  tags: string[];
}

export interface SchemesListResponse {
  schemes: SchemeEntry[];
  total: number;
}

export interface PolicyStatsResponse {
  totalSchemes: number;
  sectors: string[];
  ministries: string[];
  totalChunks: number;
  lastUpdated: string;
}

export interface VerifyClaimBody {
  /** The claim or statement to verify against policy documents */
  statement: string;
}

export interface VerifyClaimResponse {
  /** SUPPORTED | NOT_FOUND | PARTIALLY_TRUE | FALSE */
  verdict: string;
  explanation: string;
  corrections?: string;
  sources: CitedChunk[];
  confidenceScore: number;
}

export interface PolicyImpactBody {
  annualIncome: string;
  landHolding?: string;
  category: string;
  state?: string;
  occupation: string;
  age?: string;
}

export interface EligibleScheme {
  name: string;
  sector: string;
  reason: string;
  estimatedBenefit?: string;
  matchScore: number;
  actionRequired?: string;
}

export interface PolicyImpactResponse {
  eligibleSchemes: EligibleScheme[];
  totalEstimatedBenefit?: string;
  confidenceScore: number;
}

export interface SectorCount {
  sector: string;
  count: number;
}

export interface TrendingQuery {
  query: string;
  count: number;
}

export interface LatencyMetric {
  stage: string;
  ms: number;
}

export interface RetrievalMetric {
  name: string;
  value: number;
  description: string;
}

export interface PolicyAnalyticsResponse {
  totalQueries: number;
  avgResponseMs?: number;
  sectorBreakdown: SectorCount[];
  trendingQueries: TrendingQuery[];
  latencyMetrics: LatencyMetric[];
  retrieval?: RetrievalMetric[];
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateGeminiConversationBody {
  title: string;
}

export interface SendGeminiMessageBody {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GenerateGeminiImageBody {
  prompt: string;
}

export interface GenerateGeminiImageResponse {
  b64_json: string;
  mimeType: string;
}

export interface GeminiError {
  error: string;
}

export type ListSchemesParams = {
  /**
   * Filter by sector (agriculture, health, education, etc.)
   */
  sector?: string;
  /**
   * Filter by ministry name
   */
  ministry?: string;
};